pyproject.toml
[project]
name = "a2a-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "a2a-sdk[http-server,postgresql]>=0.3.22",
    "asyncpg>=0.31.0",
    "boto3>=1.42.19",
    "cairosvg>=2.8.2",
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "uvicorn>=0.40.0",
]
